Stacking Ensemble Model - Hyperparameters optimization
As part of the Final Year Project developed at the University of Westminster, a Stacking Ensemble Neural Network has been developed. We use the Sherpa library to tune the hyperparameters of the Ensemble model. The GPyOpt Bayesian algorithm has to be wrapped to be usable by the Sherpa library. The compiled neural network needs to run within a loop to allow the call-backs of the model to assess the validation loss after training. A list of Sherpa data structure items representing the grid that will be searched needs to be set. The Bayesian optimization algorithm needs to incorporate the parameters and the algorithm itself. Finally, the Bayesian Algorithm needs to be fitted with the data to start optimizing the model.